refactoring paths
