+++ embiggen/__init__.py
@@ -7,15 +7,15 @@
 
 """
 from .graph import Graph, GraphFactory
-from .text_encoder import TextEncoder
+from .embedder.word2vec.text_encoder import TextEncoder
 from .utils.tf_utils import TFUtilities
-from .w2v.cbow_list_batcher import CBOWListBatcher
-from .w2v.skip_gram_batcher import SkipGramBatcher
-from .word2vec import ContinuousBagOfWordsWord2Vec
-from .word2vec import SkipGramWord2Vec
+from .embedder.word2vec.cbow import Cbow
+from .embedder.word2vec.skipgram import SkipGram
+from .embedder.glove import GloVeModel
+from .embedder.glove import CooccurrenceEncoder
 from .embiggen import Embiggen
 
 __all__ = [
     "Graph", "GraphFactory", "Embiggen", "TextEncoder",
-    "CBOWListBatcher",  "ContinuousBagOfWordsWord2Vec", "SkipGramWord2Vec", "SkipGramBatcher",
+    "Cbow", "SkipGram","GloVeModel", "CooccurrenceEncoder"
 ]
