+++ embiggen/embedder/glove/__init__.py
@@ -1,0 +1,5 @@
+from .glove import GloVeModel
+from .coocurrence_encoder import CooccurrenceEncoder
+
+
+__all__ = ["GloVeModel", "CooccurrenceEncoder"]
